// src/app/api/swagger/models/api-trade-state-enum.ts
/* tslint:disable */
/* eslint-disable */
export enum ApiTradeStateEnum {
  Draft = 1,
  Removed = 5,
  Publishing = 10,
  FillingApplications = 15,
  AgreementConditions = 20,
  NotTookPlace = 25,
  Canceled = 30,
  ClosedWithoutDeals = 35,
  DealSigning = 40,
  DealSigned = 50,
  Archive = 55,
  Retrading = 60
}